moved Model logic to Presenter
