+++ app/src/main/AndroidManifest.xml
@@ -19,5 +19,4 @@
             </intent-filter>
         </activity>
     </application>
-
 </manifest>
