+++ app/src/main/java/com/yoshitoke/weatheringwithyou/utils/extensions.kt
@@ -3,30 +3,13 @@ package com.yoshitoke.weatheringwithyou.utils
 import java.text.SimpleDateFormat
 import java.util.*
 
-fun Int.unixTimestampToDateTimeString() : String {
+fun Int.unixTimestampToString(format: String) : String {
 
     try {
         val calendar = Calendar.getInstance()
         calendar.timeInMillis = this*1000.toLong()
 
-        val outputDateFormat = SimpleDateFormat("dd MMM, yyyy - hh:mm a", Locale.ENGLISH)
-        outputDateFormat.timeZone = TimeZone.getDefault() // user's default time zone
-        return outputDateFormat.format(calendar.time)
-
-    } catch (e: Exception) {
-        e.printStackTrace()
-    }
-
-    return this.toString()
-}
-
-fun Int.unixTimestampToTimeString() : String {
-
-    try {
-        val calendar = Calendar.getInstance()
-        calendar.timeInMillis = this*1000.toLong()
-
-        val outputDateFormat = SimpleDateFormat("hh:mm a", Locale.ENGLISH)
+        val outputDateFormat = SimpleDateFormat(format, Locale.ENGLISH)
         outputDateFormat.timeZone = TimeZone.getDefault()
         return outputDateFormat.format(calendar.time)
