+++ app/src/main/java/com/yoshitoke/weatheringwithyou/utils/network/RetrofitClient.kt
@@ -1,0 +1,28 @@
+package com.yoshitoke.weatheringwithyou.utils.network
+
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import retrofit2.Retrofit
+import retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory
+import retrofit2.converter.gson.GsonConverterFactory
+
+object RetrofitClient {
+
+    private const val BASE_URL = "https://api.openweathermap.org/data/2.5/"
+    private val logging: HttpLoggingInterceptor = HttpLoggingInterceptor().setLevel(HttpLoggingInterceptor.Level.BODY)
+    private val client: OkHttpClient = OkHttpClient.Builder().addInterceptor(logging).build()
+
+    private fun retrofit(): Retrofit {
+        return Retrofit.Builder()
+                .addConverterFactory(GsonConverterFactory.create())
+                .addCallAdapterFactory(RxJava2CallAdapterFactory.create())
+                .client(client)
+                .baseUrl(BASE_URL)
+                .build()
+    }
+
+    val weatherService: WeatherAPI by lazy {
+        retrofit().create(WeatherAPI::class.java)
+    }
+
+}
