+++ app/src/main/res/layout/activity_main.xml
@@ -9,9 +9,6 @@
     <include layout="@layout/current_conditions_layout" />
     <include layout="@layout/location_picking_layout" />
 
-    <include
-        layout="@layout/forecast_timeline_layout"
-        tools:layout_editor_absoluteX="0dp"
-        tools:layout_editor_absoluteY="0dp" />
+    <include layout="@layout/forecast_timeline_layout" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
