+++ app/src/main/res/layout/current_conditions_layout.xml
@@ -10,8 +10,8 @@
         android:id="@+id/tv_temperature"
         android:layout_width="@dimen/dp_100"
         android:layout_height="@dimen/dp_100"
-        android:layout_marginTop="160dp"
-        android:textSize="40sp"
+        android:layout_marginTop="@dimen/dp_160"
+        android:textSize="@dimen/sp_40"
         android:gravity="right"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.768"
@@ -21,23 +21,23 @@
 
     <TextView
         android:id="@+id/tv_dateTime"
-        android:layout_width="200dp"
-        android:layout_height="50dp"
-        android:layout_marginTop="4dp"
-        android:layout_marginEnd="16dp"
-        android:layout_marginRight="16dp"
-        android:textSize="15sp"
+        android:layout_width="@dimen/dp_200"
+        android:layout_height="@dimen/dp_50"
+        android:layout_marginTop="@dimen/dp_4"
+        android:layout_marginEnd="@dimen/dp_16"
+        android:layout_marginRight="@dimen/dp_16"
+        android:textSize="@dimen/sp_15"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/tv_temperature" />
 
     <TextView
         android:id="@+id/tv_condition"
-        android:layout_width="100dp"
-        android:layout_height="80dp"
-        android:layout_marginTop="168dp"
-        android:textSize="20sp"
+        android:layout_width="@dimen/dp_100"
+        android:layout_height="@dimen/dp_80"
+        android:layout_marginTop="@dimen/dp_168"
+        android:textSize="@dimen/sp_20"
         app:layout_constraintEnd_toStartOf="@+id/tv_temperature"
-        app:layout_constraintHorizontal_bias="0.496"
+        app:layout_constraintHorizontal_bias="0.5"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
