+++ app/src/main/res/layout/forecast_timeline_layout.xml
@@ -6,8 +6,8 @@
     android:layout_height="match_parent">
 
     <com.github.vipulasri.timelineview.TimelineView
-        android:layout_width="200dp"
-        android:layout_height="300dp"
+        android:layout_width="@dimen/dp_200"
+        android:layout_height="@dimen/dp_300"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
