--- app/src/main/res/values/city_list.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <string-array name="city_list">
-        <item>DaNang</item>
-        <item>HaNoi</item>
-        <item>HoChiMinh</item>
-    </string-array>
-</resources>
