+++ app/src/main/res/values/dimens.xml
@@ -2,4 +2,16 @@
 <resources>
     <dimen name="dp_100">100dp</dimen>
     <dimen name="dp_50">50dp</dimen>
+    <dimen name="dp_160">160dp</dimen>
+    <dimen name="sp_40">40sp</dimen>
+    <dimen name="dp_200">200dp</dimen>
+    <dimen name="dp_4">4dp</dimen>
+    <dimen name="dp_16">16dp</dimen>
+    <dimen name="sp_15">15sp</dimen>
+    <dimen name="dp_80">80dp</dimen>
+    <dimen name="dp_168">168dp</dimen>
+    <dimen name="sp_20">20sp</dimen>
+    <dimen name="dp_300">300dp</dimen>
+    <dimen name="dp_150">150dp</dimen>
+    <dimen name="dp_24">24dp</dimen>
 </resources>
